Added Cache to SCSS and LESS plugins.

+++ src/Assetter/Plugin/LeafoScssPhpPlugin.php
@@ -14,13 +14,51 @@ use Requtize\Assetter\PluginInterface;
 
 class LeafoScssPhpPlugin implements PluginInterface
 {
-    protected $filesRoot = null;
+    const CACHE_FILENAME = 'assetter.leafo-scss.serialize';
 
-    public function __construct($filesRoot)
+    protected $filesRoot;
+    protected $cacheDir;
+    protected $cacheData = [];
+    protected $cacheNeedRefresh = false;
+
+    public function __construct($filesRoot, $cacheDir = null)
     {
         $this->filesRoot = $filesRoot;
+        $this->setCacheDir($cacheDir ?: __DIR__.'/../../../cache');
         $this->scss = new scssc;
         $this->scss->setFormatter('scss_formatter_compressed');
+    }
+
+    public function __destruct()
+    {
+        if($this->cacheNeedRefresh)
+        {
+            file_put_contents($this->cacheDir.'/'.self::CACHE_FILENAME, serialize($this->cacheData));
+        }
+    }
+
+    public function setCacheDir($cacheDir)
+    {
+        $this->cacheDir = $cacheDir;
+
+        if(is_dir($this->cacheDir) === false)
+        {
+            mkdir($this->cacheDir, 0777, true);
+        }
+
+        if(is_file($this->cacheDir.'/'.self::CACHE_FILENAME))
+        {
+            $data = file_get_contents($this->cacheDir.'/'.self::CACHE_FILENAME);
+
+            @ $unserialized = (array) unserialize($data);
+
+            if(is_array($unserialized))
+            {
+                $this->cacheData = $unserialized;
+            }
+        }
+
+        return $this;
     }
 
     public function register(Assetter $assetter)
@@ -45,12 +83,27 @@ class LeafoScssPhpPlugin implements PluginInterface
 
     public function compile($filepath)
     {
-        $filepathNew = str_replace('.scss', '.css', $filepath);
+        $filepathRoot = $this->filesRoot.$filepath;
+        $filepathNew  = str_replace('.scss', '.css', $filepath);
 
-        $css = $this->scss->compile(file_get_contents($this->filesRoot.$filepath));
+        if($this->isFileFresh($filepathRoot) === false)
+        {
+            $css = $this->scss->compile(file_get_contents($filepathRoot));
 
-        file_put_contents($this->filesRoot.$filepathNew, $css);
+            file_put_contents($this->filesRoot.$filepathNew, $css);
+
+            $this->cacheData[$filepathRoot] = filemtime($filepathRoot);
+            $this->cacheNeedRefresh = true;
+        }
 
         return $filepathNew;
+    }
+
+    public function isFileFresh($filepath)
+    {
+        if(isset($this->cacheData[$filepath]) === false)
+            return false;
+
+        return $this->cacheData[$filepath] >= filemtime($filepath);
     }
 }
